+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added bootstrap templates
 - [codice-1] add authorization
 - [codice-2] add repos
+- [codice-9] lis repos